Made separate source-code file for ClickForPixelData class
