+++ src/main.cpp
@@ -50,6 +50,7 @@ int main(int, char **)
     {
         cap >> input_frame; //Capture image from camera.
 
+        /* NEEDED FOR CONTOUR AND MOUSE EVENT TO WORK SEPARATELY */
         output_frame_contour = input_frame.clone();
 
         //Step 1: Display original image.
